Give the constant pseudo value a size
Currently value pseudo does not have size.
This create a problem pointed out by Dibyendu.
When using LLVM, calling varidic function with
constant value, there is no where to find the
size.

Linus give out two suggestions. One is give pseudo
a size. The other one is the push instruction.
This is the implemnation of the first suggestion.

The model is actual very simple. The pseudo is
exactly as before if you are not looking at the size.
There is a size at create time, which tag alone with
it.

I might not get all the type and cast of the constant
right, we can fix it later if we found some case I did
not cover.

Testign done:

- Test suite passed.
- Kernel compile get the exact same output as the RC5
  at similar time.

Reported-by: Dibyendu Majumdar <mobile@majumdar.org.uk>
Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Christopher Li <sparse@chrisli.org>