llvm: give names easier to debug
LLVM automatically add an numeric suffix for names
automatically created. So, if intermediate names must
be created for a pseudo whose name was, for example, "%R4",
these new names will be "%R41", "%R42".

This is quite annoying because we can't make the distinction
between these names and the original names, (maybe of some other
pseudos whose names were "%R41" & "%R42).

Change this by adding a "." at the end of each name, as this will
then allow to see what the original name was.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>