cast: make [u]intptr_ctype alias of [s]size_t_ctype
When processing integers to/from pointers, we would like to have
an integer type which has the same size as a pointer.

Currently, it's always the case for [s]size_t but it's preferable
to have a specific type which will always offer this guarantee,
like [u]intptr_t.

Fix this lazily by defining [u]intptr_ctype to [s]size_t_ctype.

Note: this intptr_t is just internal to sparse and can be different
      from the type choosen by the libc/platform.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>