cast: temporary simplify handling cast to/from void*
Currently pointer casts from/to void are treated as casts
to/from integer. The rationale being that a void pointer can't
anyway be dereferenced.

Allow to continue to do so by using a new 'machine type' for void
pointers which then allow to select the right type of cast:
OP_PTRCAST or OP_[S]CAST.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>