cast: add support for -Wpointer-to-int-cast
It's relatively common to cast a pointer to an unsigned long,
for example to make some bit operations.
It's much less sensical to cast a pointer to an integer smaller
(or bigger) than a pointer is.

So, emit a diagnostic for this, under the control of a new
warning flag: -Wpointer-to-int-cast.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>