cast: add tests for warnings issued by sparse -v
The sparse command (aka the 'checker') do a number of additional
checks when used with the -v flag (I strongly believes that this
option is rarely used but let me not disgress about it here).
One of these additional checks is about casts.

Add some testcases in order to catch any problems here.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>