db/kernel.return_fixes: add dma_fence_wait_timeout()
The dma_fence_wait_timeout() returns either a negative error code or a
positive value which represents the time remaining in jiffies.  Smatch was
saying that it returned numbers as low as s64min and it all becomes
confusing if you are expecting to hold the negative return codes in an
integer type.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>