validation/sm_null_deref.c: update test
This test has been a constant source of false positives...  These days to
cut down on the false positives, we only warn about the dereference if the
pointer was initialized to NULL and not if it was set later.

Maybe some day we will turn on stricter checking again, but for now let's
make the test match the code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>