comparison: revert modification_hook_late change
I recently changed this and I did record the reason.  There was a good
reason for it, though, but I forget what.  Meanwhile the change breaks the
validation tests so I have to choose between which bugs I prefer.  And I
prefer old bugs to new bugs undocumented bugs as in "no new regressions".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>