buf_size: fix validation/sm_buf_size8.c
This validation test is taking a struct where the last member is an embeded
struct.  We're allocating extra space for the struct with malloc like this

	foo = malloc(sizeof(*foo) + 100);

The extra space is counted as belonging to the last member of the struct.
I guess the reason for this is because the embedded struct ends in a
variable length array.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>