extra: make dereferences limit to valid pointer range again
The story here is that originally dereferencing set the variable to the
valid pointer range.  But then we had the problem that there were some
pointers were already set to a sub-range of the valid pointer range and
we were just overwriting that information.  So then I changed it to say
that if we know the pointer is non-NULL then dereferencing doesn't do
anything but that broke some of the validation tests.

Now we use rl_intersection() and hopefully everything is happy.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>