validation/sm_compare12.c: update to reflect latest code
We can generally assume that incrementing a variable, doesn't make it
overflow.  Most often that's the case, and pretending it's always the case
even when we're not sure silences a bunch of false positives.

Smatch doesn't handle this totally correctly.  If we increment the lower
variable one time, it should say that we first go to <= then increment
one more time and it goes to <none>.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>