db: move the !mem_db closer to where it's needed
We don't need to turn of the whole DB if there is no mem_db.  I don't think
this makes a difference very often since we should always have the mem_db.
The reason I noticed it was because I had an old smatch_data/ directory and
this check created extra warning messages that didn't help.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>