db: Use smatch_data/db/<project>.return_fixes on the mem DB
One bug which I have had several times is that I set the return_fixes but
I forget that actually the code is always parsing the function inline
instead of querying the DB.  I've fixed this bug by swapping the
return_fixes on inline functions as well.  This means I could probably
delete the code for the that from create_db.sh but I haven't yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>