fix-return: remove special case for single return
During the linearization of a function, returns are directly
linearized as phi-sources and the exit BB contains the
corresponding phi-node and the unique OP_RET.

There is also a kind of optimization that is done if there is
only a single a return statement and thus a single phi-source:
the phi-source and the phi-node is simply ignored and the
unique value is directly used by the OP_RET instruction.

While this optimization make sense it also has some cons:
- the phi-node and the phi-source are created anyway and will
  need to be removed during cleanup.
- the corresponding optimization need to be done anyway during
  simplification
- it's only a tiny special case which save very litte.

So, keep things simple and generic and leave this sort of
simplification for the cleanup/simplification phase.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>