type_val: make is_uncasted_function() stricter
In type_val we're trying to record every time a value is set.  There are
two possible mistakes, one is recording places where it wasn't really set
and the other is missing places where it does get set.

The is_uncasted_function() is supposed to record things like:

	struct my_struct *p = kmalloc();

It records that all the struct members are set to random data.  But if you
have a function that returns a my_struct pointer that isn't recorded
because we assume that is recorded within the function itself.  (This
probably isn't always true in the current code, but that's a separate bug).

The problem here is that basically any call assignment was ignored.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>