user_data2: change how conditions are handled to silence underflow warnings
What I was looking at here is this:

	if (untrusted_int < trusted_ulong)
		my_array[untrusted_int] = foo;

The "untrusted_int" here would be type promoted to unsigned long.  If it
were negative, that would be a huge positive value.  Probably the trusted
value is the size of my_array[] and not something larger than INT_MAX.
In the original code, Smatch would say, "we don't know the value of
trusted_ulong so it could be larger than INT_MAX", but now we just
assume that's not true.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>