testsuite: remove useless test for loop-linearization
This testcase was added a bit too quickly in order to have
minimal testing of loop's linearization.

However, such test just comparing the raw output of test-linearize
is a big PITA because it's so sensible to things like pseudos' name
themselves depending very much on details about the linearization
and simplification. Also, this test didn't really tested anything,
it only allowed to track changes.

Remove it as it has no testing value.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>