param_set: don't ignore "fake" modifications
The thinking here was that if we have statement like:

	param = b;

Then it creates fake modifications like:

	param->foo = fake_call();

But we don't need to record that param->foo was modified since we already
know that "param" was modified and that's enough.  The bug is that there
are other types of fake calls and those were getting ignored as well.  Code
like:

	foo = (({whatever ? a : b}));

Also results in a fake call.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>