debug: add __smatch_exit() which halts parsing
This is useful for when you are trying to debug a crash and you want to
see which line it's on.  You can move the exit statement around to see if
it crashes before the exit or afterward.  Also it speeds up parsing if you
can exit towards the beginning of a file.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>