user_data2: use the extra_nomod_hooks
user_data2 is always a subset of smatch_extra.c.  I had thought I could
just track when we set new user data and then when get_user_rl() is called,
I would take the intersection of smatch_extra.c and the user_data state.

But actually the extra state sometimes changes behind our backs so that
doesn't work.  We have to be able to handle if statements.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>