user_data2, db: fix changing the returned states to $->foo->bar correctly
The get_param_name() works for parameters, because parameters are all
symbol expressions, but returned expressions can be dereferences like
foo->bar.  The original code was taking foo->baz and change it to "$->baz"
but it should have ignored it.  And it takes things like "foo->bar->xxx"
and changes it to "$->bar->xxx" instead of "$->xxx".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>