teach sparse about '-fmax-warnings'
Currently, sparse stops to emit warnings after the 100th.
This is, of course, to avoid to throw lots of warnings if the
source file has lots of problems but sometimes we want to see
all warnings & errors.

For example, it's useful when testing a new version of
sparse on the kernel since it has a few files with more
than 100 warnings (and so a bad change will be undetected
if it affects warnings only after the 100th).

Let parse the '-fmax-warnings=N' option and limit the
number of warnings accordingly.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>