teach sparse about '-o <file>'
Sparse knows about the '-o' option, parses it but does
nothing with it.

Change this by redirecting stdout to <file> unless <file>
is '-' since sparse (the lib) outputs to stdout by default.
But ignore this flag when sparse is used purely as an checker
since in this case it's not supposed to output to stdout
and would create undesired empty file, possibly erasing the
result of the compiler if one is used before sparse.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>