db: add function_signature to the INTERNAL entries
The problem here is that you can have a two function pointers like this
and:

	(struct struct_name)->function_name(int, char *);
	(struct struct_name)->function_name(char *);

The struct name and the member names are the same (they are defined in the
.c file instead of the .h file) but they take different arguments.  This
happens because code is copy and pasted so the names are identical but
later the parameters are updated.  We can filter based on the arguments.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>