estate: estate_is_empty() should return false if the state is NULL
An empty is estate is a specific thing that means that the condition is
impossible.  It's for code like:  "foo = NULL;  if (foo) { ...".  On the
other hand if the estate is NULL that means the variable could be anything
those are two totally different things.

I just added a caller smatch_extra where we check if the estate is empty
when handling comparisons with zero and it broke because of this bug.
There are two other callers but they were only using it on non-NULL states.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>