idf: compute the iterated dominance frontier
Conversion of some IR into SSA form requires to place so called
phi-nodes where different paths for some value meet. It's important,
of course to place these correctly, but it's also important to try
to minimize the number of these phi-nodes. Several algorithms exist
for the placing of these phi-nodes but it can be shown that, for
each variable, it is suffisant to place their phi-nodes on the
dominance frontier of the nodes defining this variable and since
phi-nodes themselves give a new definition, to have the minimal
number of phi-nodes, these phi-nodes must be paced on the iterated
dominance frontier (the transitive closure of the dominance frontier
of the initial defining nodes).

This patch implement what's needed to compute the iterated dominance
frontier of a set of nodes.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>