extra: fix comparisons with zero (more)
While I was fixing the bug that estate_is_empty() treats NULL states as
empty I became suspicious of this code and re-wrote it.  My worry is that
instead along side the smatch_extra state we should also be considering
information stored in the database for the true path.

In the current code, it assumes that because the estate is NULL then that
means we must take the whole range.  But instead we should look at the DB
and perhaps we can tell from the type that the range is limited.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>