helper: re-order symbol additions in expr_to_chunk()
The problem is say you have code like:

	if (foo + bar > 5) {

we know the value of "foo + bar" but say we want to know the value of
"bar + foo"?  Smatch doesn't know those two are the same.  Most code uses
the same ordering consistently but what often causes the problem is code
like this:

	if (foo > 5)
		return -EINVAL;
	if (bar > 5 - foo)
		bar =  5 - foo;

For code like that, Smatch picks the order of "foo + bar" or "bar + foo"
so it might not be consistent with how the programmer uses it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>