defer initialization of bb::context
Currently bb::context is intialized (at -1) when the
basic block is allocated.

But this field:
1) is only used when using the sparse tools;
2) when used, it's only quite late in the process;
3) this early initialization prevents us to reuse the space
   of this field for another purpose, earlier in the process.

Change this by initializing this field much later, by the sparse tool
itself, just before needing it.

The real motivation being, of course, to be able to reuse the space
for some other upcoming field.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>