helper: improve expr_to_str() for sizeof()
Handling sizeof() is a problem.  Ideally it expr->cast_type would always
be true, but sometimes the expression gets evaluated by Sparse and becomes
just another number.

Smatch really wants to know about sizeof() calls because sizeof(foo) is not
just a number, it indicates a relationship between then number and foo.
I've tried to explain this to the Sparse people but they say I am
wrong/crazy.

Anyway, if all we have is the number, then print the number.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>