constraints: add constraints framework
This code is not useful at present.

In the current code, we mostly record that a variable is within 0-65 range.
If we have a condition like:

	if (foo >= bar->baz)
		return -EINVAL;

Then we try to look up bar->baz in the type_value table, but if we can't
find it then we just mark "foo" as capped.  But unfortunately, just saying
&capped is not fine grained enough.  For example, if we do:

	p = kmalloc(foo, GFP_KERNEL);

If p is non-NULL that means that foo is limited to less than 4MB so foo is
capped.  Lot's of things can cap foo but they don't necessarily mean it's
safe to use.

So now I'm introducing smatch_constraints.c which is more specific.  It
tracks that "if 'idx' is less than 'foo->limit' then we can use it as
index into the foo->data[] array."  I worried that recording all the limits
in the database as a string would require too much space so I've created a
list of all the limits and they get assigned an ID.  The ID is preserved
when you rebuild the database on new kernel source.  Then we record that
the condition is met for param 2 or whatever and that it tracked across
function boundaries.

This code doesn't isn't really useful yet and hasn't found any bugs.  But
I hope that long term it is the way forward.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>