db: try to create a link from data to function (through function pointers)
Most function pointer calls let you specify a data argument.  For example
you could have something like this:

	foo->func = my_func;
	foo->data = my_data;

	foo->func(foo->data, blah, blah);

The goal of this is to track it back so we can link a function with a
specific argument.  The first thing I want to track is just the type of
argument so that I can from within the function if a void pointer is just
a specific struct type casted to void.  It works imperfectly, but hopefully
it's a platform which can be fixed later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>