type_val: try to use the function pointer arg link for ignoring void parameters
One of the problems that I have in Smatch is that the kernel uses a lot
of function pointers which look like this:

void frob(void *p)
{
	struct my_type *foo = p;


Then in smatch_type_val.c we say "Huh...  'p' could be anything so let's
mark my_type->value as a complete unknown."  But if we knew we were just
casting a my_type struct to void and then back we could ignore the "p"
assignment.

Btw, I've played around with just ignoring every void pointer assignment
like this, but it leads to false positives and didn't really help find
any new bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>