ranges: handle truncated ranges better (with show_rl() and parsing data)
The bug here was that we were passing a lot of commands to a function
and each command is an int so it looks like:

	cmd = 1,4,5,6,2432,33453,345345,35434345,...

And show_rl() will only print 256 characters so the last commands were
silently truncated away.  Then when we read the database we say, "That
function will never return when we pass it that command" so a lot of
paths get marked as impossible when they shouldn't be.

My fix is to boost the buffer to 512 characters and if we don't have
enough space to print the whole range list then put a '+' at the end of
the string.  Whenever we see that it means that any higher value might be
valid.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>