function_ptrs: fix is_local_symbol()
We want to test if a function pointer is local to a function so we can
just test the MOD_TOPLEVEL flag.  The problem in the original code is that
it assumed that function pointers were declared at the start of the
function instead of later in a more limited scope.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>