db: record common function calls but don't select the info
The problem with common functions is that when you're trying to parse them
and they're called hundreds of times then Smatch just hangs trying to parse
all the information.  And it's likely useless anyway...

In the original code, I just didn't record the common functions in the
database.  But it is useful information if you're trying to build a
call_tree, so with this patch, the calls are recorded but we just don't
select the information when we're parsing the function.

This patch is pretty crap.  The list of common functions is stored in
smatch_data/kernel.common_functions.  The parser for that can't handle
function pointers so I end up handling function pointers with the old
way (I detect function pointers by looking for a space in the name).
Anyway, hopefully I will re-write this code soon.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>