extra: delete match_untracked_array()
I'm not sure using smatch_untracked is appropriate here.  Say we pass a
pointer "p" to a function assigned "foo->bar = p;" then we mark that
as untracked.

And sure, we *might* lose track of it, but mostly likely it's just saved
for later and not changed.  So it's better to assume that we still know the
value of "p->x;" or whatever.  This fixes a false postive:

drivers/media/usb/pulse8-cec/pulse8-cec.c:685 pulse8_connect() error: 'pulse8->adap' dereferencing possible ERR_PTR()

The check uses smatch_extra to see if a variable can possibly be in the
error pointer range, so in the original code smatch_extra says "I don't
know, it could be anything".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>