db/kernel.return_fixes: hack scnprintf() return
The issue here is that we really want to specify that the return can't be
negative.

Part of the story is that Smatch has a hard time parsing this because we
currently either store a math like: s32min-s32max[$1 - 1] or a
comparison: 0-s32max[<$1].  And in theory a math aught to translate into
a comparison but it doesn't necessarily.  Also smatch comparisons need
some work because it should probably be "expr vs expr" instead of
"str vs str".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>