comparison: fix a bug with unsigned/signed impossible returned comparisons
The problem is that we're saying that "foo is UNSIGNED_LT bar" and then
the db doesn't record signed/unsigned information.  Then there is probably
another bug that filter_comparison doesn't handle signed/unsigned correct?
I'm not positive if this a bug or a feature...

Anyway, I feel like this patch is probably the correct thing.  We can
normally assume that the comparisons and the DB comparisons are going to
use the same signedness...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>