*new* check_uninitialized: warn about uninitialized variables
This check is a cross function check for uninitialized variables.  GCC
basically assumes if you pass a variable as a pointer then, it comes back
initialized.  GCC generally misses a lot of uninitialized variable bugs
which you would expect it to catch and I don't know the reason.

This test produces about 1.3k warnings on my system.

The most common false positives are because:

*) We're reading from the hardware and we know it's not going to fail so we
   don't bother checking the error paths.  I've added a
   smatch_data/kernel.ignore_uninitialized_param file to handle this case.
   But it doesn't work for function pointer calls.

*) Inconsistent checking between "if (ret) {" and "if (ret < 0) {".  The
   kernel mostly uses negative error codes but most code treat all
   non-zeros as errors.  There are a couple subsystems where maintainers
   prefer "if (ret < 0) {", like sound/ and parts of net/, but you end up
   with a mix.  So the function is saying, I return non-zero on error, but
   the caller only checks for negatives.

   Of course, Smatch is trying to track the actual error codes but there
   are many ways this breaks down.  For example, if the error code is set
   in another thread, if the function is "too hairy" to parse or if it's
   written in assembly.

Those are the big ones I can think of...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>