param_used: record when a container_of() a param is used
The kernel uses the container_of() macro a lot.  This is the first
step towards supporting that.  It inserts something with a key like this
into call implies "$(-346)->foo" to say that we take the container_of
the parameter, and use the ->foo member.  The "346" is the offset in bytes.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>