db: use the struct_member_callbacks to save container_of() information
We first look up to see if the container_of() is used.  Then we try to
find the container.  Then we use the existing struct_member_callbacks to
print the information.

The downside of this approach is that you need to build the DB twice at a
minimum to store any container_of() information.  But we already have that
requirement to build the DB twice or more for other thins.  Really the idea
is that you build the DB every morning...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>