db, struct_assignment: handle container_of() assignments
We save the container_of() information in the caller_info table like this
"$(-346)->foo".  I at first thought I would have to write some code to
map that to the variable names in the function but it turns out that
smatch_extra already "handles" them.  It save the caller_info as:
"foo(-346)->foo = 100", which is nonsense, but at least it can't possibly
conflict with any real variables so it's as good as any format I can think
of off the top of my head.

Then the other thing is that we really want to preserve the implications
and that means cloning the sm_states.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>