db: change how common functions are handled
There are some functions which are called too many times to merge all their
caller_info together.  At first I didn't save the info, then I saved it but
I had a file with a list of functions to not look up.  But now I've
created a new table which mirrors caller_info but it's only for functions
which are called too many times to track.

For now it just has a fake call with no information but, later, I probably
will want to load it with merged data if possible.

I left the code to not load the too common functions, but I changed the
script to not add functions to that file automatically.  I think I will
probably use it again so that we can remove some genuinely uninteresting
function calls from the database.  The database is getting too large.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>