db: split caller_info and sql into separate output file
The DB is getting too huge and it takes too long to create.  One easy win
is to split the caller_info into a separate file from the other sql so that
we don't have to read the whole massive file over and over to find the
information we want.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>