db: handle returning block statements better
This is part of handling container_of() better.  Say you have code like:

	return container_of(...);

The container_of() macro is one of those block statements like so:

#define container_of(ptr, type, member) ({              \
	void *__mptr = (void *)(ptr);						\
	((type *)(__mptr - offsetof(type, member))); })

Normally when we handle that sort of thing, then we just consider the
last statement in the block statement.  And we should do the same thing
with return statements as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>