db: handle functions that return a container_of()
It's pretty common to have functions which are just a wrapper around
container_of() like so:

static struct my_foo *bar_to_foo(struct my_bar *p)
{
	return container_of(p, struct my_foo, p);
}

This patch adds support for that.  We return a CONTAINER with the param
and the offset.  Then when we get it, we save that it's a param_used and
if we have any states from caller_info then we copy those to the
appropriate variables.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>