flow: parse more inline functions
Sparse gives Smatch a list of functions, but leaves out all the inline
functions.  Smatch handles this by tracking the inline functions which are
called and then parsing them after the end of the first function to call
them.

I had thought this would work, but it turns out that depending on the
config there can be syscalls which are inline functions.  The syscalls are
called by the compat version of the syscall, but that happens before the
real implementation is defined so it means we can't parse them.  I don't
know how this works at all, but presumably it has something to do with
linker magic.

Anyway, syscalls are very important and we should parse them.  I at first
modified Sparse to give Smatch every function instead of leaving out the
inline functions.  The problem was that ended up creating hundreds of gigs
of useless cross function data.  So then I modified smatch_flow.c to only
parse the new inline functions if they are in a .c file.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>