param_used/states: introduce __get_state() for silently looking up states
I'm becoming more conflicted about smatch_param_used.c.  It tracks whenever
we look up a state and marks it as used.  Then if something isn't used,
that means we don't need to pass the information to that function.

The problem here is that we were looking up stuff to just do bookkeeping
and not because we were really interested in it.  So it gets marked as
used when it's not.  We end up tracking a lot of information we don't
need.

I've introduced a new __get_state() function which like get_state() but
only for bookkeeping purposes.  The plan is to use it on a case by case
basis when the cross function DB gets too large.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>