sparse: document that -Wbitwise is default
In commit commit 02a886bfa3d9 ("Introduce keyword driven
attribute parsing", 2007-03-08) the -Wbitwise keyword was broken and did
not actually work. Instead, bitwise checks were always enabled.

This was fixed by commit commit 0dfda0d1f0fe ("make -Wbitwise
operational again", 2017-02-18) which allowed -Wbitwise and -Wno-bitwise
to work as expected.

However, since -Wbitwise was enabled for so long, that commit changed
the default of -Wbitwise to be enabled, rather than the documented
"disabled".

Fix the documentation to match, and explain that -Wbitwise is enabled by
default and must be disabled.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>
Acked-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Signed-off-by: Christopher Li <sparse@chrisli.org>