buf_size: don't store obvious sizes in the DB
Say a function takes a struct foo pointer, it's waste of time to store that
we're passing a struct foo size object.  That's assumed from the type.

Also the old behavior caused a lot of false positives when the struct
changed to a different size and it took a while for the size information to
propagate through the call tree.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>