var_sym: fix handling of arrays
The breakage was detected with the validation test suite.

I wrote this thinking that we could take a short cut if there was just one
sym involved.  But in an array like "array[2]" there is only one sym but
it's not a variable so we can't take the short cut.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>